scripts/az_to_gcp_cpsecret/requirements.txt
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
google-cloud-secret-manager==2.18.1

